MNT: Test against different pyvista versions (#110)
* Test against different pyvista versions

* Update job name

* Update job name

* DeprecationWarning as errors

* Fix

* Add infos

* Update

* Update

+++ pyvistaqt/plotting.py
@@ -271,7 +271,7 @@ class QtInteractor(QVTKRenderWindowInteractor, BasePlotter):
             # Modified() and upstream objects won't be updated.  This
             # ensures the render window stays updated without consuming too
             # many resources.
-            twait = (auto_update ** -1) * 1000.0
+            twait = int((auto_update ** -1) * 1000.0)
             self.render_timer.timeout.connect(self.render)
             self.render_timer.start(twait)
